Move `PeerInfo` and `Persistence` to a separate module
